# Please remove the below "#" ahead of kelude bug info, the first one will be the git commit subject
Bugfix: handle null correctly

+++ src/main/java/com/aliyun/odps/jdbc/utils/transformer/DoubleTransformer.java
@@ -7,7 +7,7 @@ public class DoubleTransformer extends AbstractTransformer {
   @Override
   public Object transform(Object o, String charset) throws SQLException {
     if (o == null) {
-      return 0;
+      return (double) 0;
     }
 
     if (Number.class.isInstance(o)) {
