+++ src/main/java/com/aliyun/odps/jdbc/utils/transformer/IntTransformer.java
@@ -7,7 +7,7 @@ public class IntTransformer extends AbstractTransformer {
   @Override
   public Object transform(Object o, String charset) throws SQLException {
     if (o == null) {
-      return 0;
+      return (int) 0;
     }
 
     if (Number.class.isInstance(o)) {
