+++ src/main/java/com/aliyun/odps/jdbc/utils/transformer/ShortTransformer.java
@@ -7,7 +7,7 @@ public class ShortTransformer extends AbstractTransformer {
   @Override
   public Object transform(Object o, String charset) throws SQLException {
     if (o == null) {
-      return 0;
+      return (short) 0;
     }
 
     if (Number.class.isInstance(o)) {
